refactor(api): migrate errors used by authentication usecases and services
