+++ api/lib/domain/usecases/authenticate-external-user.js
@@ -1,14 +1,16 @@
 import {
-  MissingOrInvalidCredentialsError,
   UserNotFoundError,
-  PasswordNotMatching,
   UnexpectedUserAccountError,
   UserAlreadyExistsWithAuthenticationMethodError,
-  UserShouldChangePasswordError,
 } from '../errors.js';
 
 import { AuthenticationMethod } from '../models/AuthenticationMethod.js';
 import { NON_OIDC_IDENTITY_PROVIDERS } from '../constants/identity-providers.js';
+import {
+  MissingOrInvalidCredentialsError,
+  PasswordNotMatching,
+  UserShouldChangePasswordError,
+} from '../../../src/access/shared/domain/errors.js';
 
 async function authenticateExternalUser({
   username,
