+++ api/tests/integration/application/error-manager_test.js
@@ -1,6 +1,10 @@
 import { expect, HttpTestServer, sinon } from '../../test-helper.js';
 import * as DomainErrors from '../../../lib/domain/errors.js';
-import { ForbiddenAccess, EntityValidationError, CsvImportError } from '../../../src/shared/domain/errors.js';
+import { CsvImportError, EntityValidationError, ForbiddenAccess } from '../../../src/shared/domain/errors.js';
+import {
+  MissingOrInvalidCredentialsError,
+  UserShouldChangePasswordError,
+} from '../../../src/access/shared/domain/errors.js';
 
 describe('Integration | API | Controller Error', function () {
   let server;
@@ -806,7 +810,7 @@ describe('Integration | API | Controller Error', function () {
     const UNAUTHORIZED_ERROR = 401;
 
     it('responds Unauthorized when a MissingOrInvalidCredentialsError error occurs', async function () {
-      routeHandler.throws(new DomainErrors.MissingOrInvalidCredentialsError());
+      routeHandler.throws(new MissingOrInvalidCredentialsError());
       const response = await server.requestObject(request);
 
       expect(response.statusCode).to.equal(UNAUTHORIZED_ERROR);
@@ -832,9 +836,7 @@ describe('Integration | API | Controller Error', function () {
     });
 
     it('responds Unauthorized when a UserShouldChangePasswordError error occurs', async function () {
-      routeHandler.throws(
-        new DomainErrors.UserShouldChangePasswordError('Erreur, vous devez changer votre mot de passe.'),
-      );
+      routeHandler.throws(new UserShouldChangePasswordError('Erreur, vous devez changer votre mot de passe.'));
       const response = await server.requestObject(request);
 
       expect(response.statusCode).to.equal(UNAUTHORIZED_ERROR);
