+++ api/tests/shared/unit/domain/services/encryption-service_test.js
@@ -3,7 +3,8 @@ import bcrypt from 'bcrypt';
 import { catchErr, expect } from '../../../../test-helper.js';
 
 import * as encryptionService from '../../../../../src/shared/domain/services/encryption-service.js';
-import { PasswordNotMatching } from '../../../../../lib/domain/errors.js';
+
+import { PasswordNotMatching } from '../../../../../src/access/shared/domain/errors.js';
 
 describe('Unit | Shared | Domain | Services | Encryption', function () {
   describe('#checkPassword', function () {
