+++ api/tests/unit/domain/usecases/authenticate-external-user_test.js
@@ -2,16 +2,18 @@ import { expect, sinon, domainBuilder, catchErr } from '../../../test-helper.js'
 import { authenticateExternalUser } from '../../../../lib/domain/usecases/authenticate-external-user.js';
 
 import {
-  MissingOrInvalidCredentialsError,
   UserNotFoundError,
-  PasswordNotMatching,
   UnexpectedUserAccountError,
-  UserShouldChangePasswordError,
   UserAlreadyExistsWithAuthenticationMethodError,
 } from '../../../../lib/domain/errors.js';
 
 import { AuthenticationMethod } from '../../../../lib/domain/models/AuthenticationMethod.js';
 import { NON_OIDC_IDENTITY_PROVIDERS } from '../../../../lib/domain/constants/identity-providers.js';
+import {
+  MissingOrInvalidCredentialsError,
+  PasswordNotMatching,
+  UserShouldChangePasswordError,
+} from '../../../../src/access/shared/domain/errors.js';
 
 describe('Unit | Application | UseCase | authenticate-external-user', function () {
   let tokenService;
